requirements.txt
Flask
Requests
Flake8
pytest
PyYAML